# product-service/swagger.yaml
openapi: 3.0.0
info:
  title: Product Service API
  version: 2.0.0

components:
  schemas:

    Product:
      type: object
      required:
        - id
        - name
        - description
        - price
        - category
        - brand
        - stock
      properties:
        id:
          type: integer
          example: 1

        name:
          type: string
          example: "iPhone 15"

        description:
          type: string
          example: "Latest Apple smartphone"

        price:
          type: number
          example: 1500

        category:
          type: string
          example: "Electronics"

        brand:
          type: string
          example: "Apple"

        stock:
          type: integer
          example: 20

        createdAt:
          type: string
          format: date-time

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        brand:
          type: string
        stock:
          type: integer
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string

paths:

  /products:
    get:
      summary: Get all products
      responses:
        '200':
          description: List of products
        '500':
          description: Server error

    post:
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              id: 1
              name: "iPhone 15"
              description: "Latest Apple smartphone"
              price: 1500
              category: "Electronics"
              brand: "Apple"
              stock: 20
      responses:
        '201':
          description: Product created
        '400':
          description: Validation error or duplicate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /products/{id}:
    get:
      summary: Get product by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Product found
        '404':
          description: Not found

    put:
      summary: Update product (partial update supported)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
            example:
              name: "iPhone 15"
              description: "Latest Apple smartphone"
              price: 1500
              category: "Electronics"
              brand: "Apple"
              stock: 20
      responses:
        '200':
          description: Updated successfully
        '400':
          description: Validation error
        '404':
          description: Not found

    delete:
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
      responses:
        '200':
          description: Deleted
        '404':
          description: Not found